+++ external_libraries/FindBaselibs.cmake
@@ -219,7 +219,22 @@ if (Baselibs_FOUND)
   #       to in the future with spack)
 
   # HDF5_LIBRARIES is a list of libraries that HDF5 needs to link to
-  set (HDF5_LIBRARIES hdf5_hl_fortran hdf5_fortran hdf5_hl hdf5 sz z m dl)
+
+  # We need to be careful here. If Baselibs was built with libaec, then
+  # the "sz" library is actually "aec sz". So we need to check if
+  # libaec is in the BASEDIR/lib directory. If it is, then we
+  # can set SZ_LIB to "aec sz", otherwise we set it to just "sz".
+  if (EXISTS ${BASEDIR}/lib/libaec.a)
+    # If we have libaec, then we use it
+    set (SZ_LIB "aec sz")
+    message(STATUS "Found libaec in BASEDIR/lib. Using aec sz for SZ_LIB.")
+  else ()
+    # If we don't have libaec, then we use just sz
+    set (SZ_LIB "sz")
+    message(STATUS "Did not find libaec in BASEDIR/lib. Using sz for SZ_LIB.")
+  endif ()
+
+  set (HDF5_LIBRARIES hdf5_hl_fortran hdf5_fortran hdf5_hl hdf5 ${SZ_LIB} z m dl)
   # Create targets
 
   # - HDF5 C