+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Add support for libaec in Baselibs
+
 ### Deprecated
 
 ## [3.62.1] - 2025-05-22